+++ scripts/examples/map/map-GeneralExample.daph
@@ -24,20 +24,16 @@ print(map(X, timesTwo));
 
 // Python Lambda Expression
 choosen_function = "x*2";
-print(map(X, choosen_function, "x", "Python_Ctypes_SysArg"));
-print(map(X, choosen_function, "x", "Python_Ctypes_sharedMem_address"));
-print(map(X, choosen_function, "x", "Python_Ctypes_sharedMem_voidPointer"));
-print(map(X, choosen_function, "x", "Python_Ctypes_sharedMem_Pointer"));
-print(map(X, choosen_function, "x", "Python_Ctypes_copy"));
-print(map(X, choosen_function, "x", "Python_Ctypes_binaryData"));
-print(map(X, choosen_function, "x", "Python_Ctypes_csv"));
+print(map(X, choosen_function, "x", "Python_SysArg"));
+print(map(X, choosen_function, "x", "Python_Shared_Mem"));
+print(map(X, choosen_function, "x", "Python_Copy"));
+print(map(X, choosen_function, "x", "Python_BinaryFile"));
+print(map(X, choosen_function, "x", "Python_CsvFile"));
 
 //Python Function
 choosen_function = "def double(x): return x*2";
-print(map(X, choosen_function, "x", "Python_Ctypes_SysArg"));
-print(map(X, choosen_function, "x", "Python_Ctypes_sharedMem_address"));
-print(map(X, choosen_function, "x", "Python_Ctypes_sharedMem_voidPointer"));
-print(map(X, choosen_function, "x", "Python_Ctypes_sharedMem_Pointer"));
-print(map(X, choosen_function, "x", "Python_Ctypes_copy"));
-print(map(X, choosen_function, "x", "Python_Ctypes_binaryData"));
-print(map(X, choosen_function, "x", "Python_Ctypes_csv"));
+print(map(X, choosen_function, "x", "Python_SysArg"));
+print(map(X, choosen_function, "x", "Python_Shared_Mem"));
+print(map(X, choosen_function, "x", "Python_Copy"));
+print(map(X, choosen_function, "x", "Python_BinaryFile"));
+print(map(X, choosen_function, "x", "Python_CsvFile"));
