+++ src/runtime/local/kernels/CMakeLists.txt
@@ -83,13 +83,11 @@ list(APPEND LIBS DataStructures IO BLAS::BLAS MLIRParser)
 set(PREFIX ${PROJECT_SOURCE_DIR}/src/runtime/local/kernels/)
 set(HEADERS_cpp_kernels
     ${PREFIX}/MatMul.h
-    ${PREFIX}/MAP_EXTERNAL/CtypesMapKernel_sharedMem_voidPointer.h
-    ${PREFIX}/MAP_EXTERNAL/CtypesMapKernel_sharedMem_address.h
-    ${PREFIX}/MAP_EXTERNAL/CtypesMapKernel_sharedMem_Pointer.h
-    ${PREFIX}/MAP_EXTERNAL/CtypesMapKernel_csv.h
-    ${PREFIX}/MAP_EXTERNAL/CtypesMapKernel_binaryData.h
-    ${PREFIX}/MAP_EXTERNAL/CtypesMapKernel_copy.h
-    ${PREFIX}/MAP_EXTERNAL/CtypesMapKernel_SysArg.h
+    ${PREFIX}/MAP_EXTERNAL/PythonMapKernel_shared_mem.h
+    ${PREFIX}/MAP_EXTERNAL/PythonMapKernel_csv.h
+    ${PREFIX}/MAP_EXTERNAL/PythonMapKernel_binaryData.h
+    ${PREFIX}/MAP_EXTERNAL/PythonMapKernel_copy.h
+    ${PREFIX}/MAP_EXTERNAL/PythonMapKernel_SysArg.h
 )
 set(SOURCES_cpp_kernels
         ${PREFIX}/MatMul.cpp
