+++ src/runtime/local/kernels/MAP_EXTERNAL/CMakeLists.txt
@@ -3,5 +3,5 @@ set(CMAKE_CXX_STANDARD 17 CACHE STRING "C++ standard to conform to")
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 cmake_minimum_required(VERSION 3.21)
 
-project(mapExternalPL_CtypesKernel)
-message(STATUS "CtypesKernel CMake Build")
+project(mapExternalPL_Kernel)
+message(STATUS "Python Kernel CMake Build")
