+++ src/runtime/local/kernels/MAP_EXTERNAL/PythonMapKernel_binaryData.h
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-#ifndef SRC_RUNTIME_LOCAL_KERNELS_MAP_BINARYDATA_BINARYDATAMAPKERNEL_H
-#define SRC_RUNTIME_LOCAL_KERNELS_MAP_BINARYDATA_BINARYDATAMAPKERNEL_H
+#ifndef SRC_RUNTIME_LOCAL_KERNELS_MAP_EXTERNAL_PYTHONMAPKERNEL_BINARYDATA_H
+#define SRC_RUNTIME_LOCAL_KERNELS_MAP_EXTERNAL_PYTHONMAPKERNEL_BINARYDATA_H
 
 #include <runtime/local/datastructures/DenseMatrix.h>
-#include <runtime/local/kernels/MAP_EXTERNAL/MapKernelUtils.h>
+#include <runtime/local/kernels/MAP_EXTERNAL/PythonMapKernelUtils.h>
 #include <Python.h>
 #include <memory>
 #include <util/PythonInterpreter.h>
@@ -27,18 +27,18 @@
 #include <cstdio>
 
 template<typename DTRes, typename DTArg>
-struct CtypesMapKernel_binaryData
+struct PythonMapKernel_binaryData
 {
     static void apply(DTRes *& res, const DTArg * arg, const char* func, const char* varName) = delete;
 };
 
 template<class DTRes, class DTArg>
-void ctypesMapKernel_binaryData(DTRes *& res, const DTArg * arg, const char* func, const char* varName) {
-    CtypesMapKernel_binaryData<DTRes,DTArg>::apply(res, arg, func, varName);
+void pythonMapKernel_binaryData(DTRes *& res, const DTArg * arg, const char* func, const char* varName) {
+    PythonMapKernel_binaryData<DTRes,DTArg>::apply(res, arg, func, varName);
 }
 
 template<typename VTRes, typename VTArg>
-struct CtypesMapKernel_binaryData<DenseMatrix<VTRes>, DenseMatrix<VTArg>> {
+struct PythonMapKernel_binaryData<DenseMatrix<VTRes>, DenseMatrix<VTArg>> {
 
     static void apply(DenseMatrix<VTRes> *& res, const DenseMatrix<VTArg> * arg, const char* func, const char* varName)
     {
@@ -55,7 +55,7 @@ struct CtypesMapKernel_binaryData<DenseMatrix<VTRes>, DenseMatrix<VTArg>> {
         output.write(reinterpret_cast<const char *>(arg->getValues()), arg->getNumRows() * arg->getNumCols() * sizeof(VTArg));
         output.close();
 
-        PyObject* pName = PyUnicode_DecodeFSDefault("CtypesMapKernel_binaryData");
+        PyObject* pName = PyUnicode_DecodeFSDefault("PythonMapKernel_binaryData");
         PyObject* pModule = PyImport_Import(pName);
         Py_XDECREF(pName);
         if (!pModule) {
@@ -110,4 +110,4 @@ struct CtypesMapKernel_binaryData<DenseMatrix<VTRes>, DenseMatrix<VTArg>> {
         PyGILState_Release(gstate);
     }
 };
-#endif //SRC_RUNTIME_LOCAL_KERNELS_MAP_BINARYDATA_BINARYDATAMAPKERNEL_H
+#endif //SRC_RUNTIME_LOCAL_KERNELS_MAP_EXTERNAL_PYTHONMAPKERNEL_BINARYDATA_H
