+++ src/runtime/local/kernels/MAP_EXTERNAL/PythonMapKernel_binaryData.py
@@ -20,17 +20,13 @@
 # Modifications Copyright 2022 The DAPHNE Consortium
 #
 # -------------------------------------------------------------
-
-import MapKernelUtils
+import PythonMapKernelUtils
 import numpy as np
-import pandas as pd
 from sympy import symbols, lambdify, sympify, Symbol
 import re
 
 def apply_map_function(input_file, output_file, rows, cols, func, varName, dtype):
-
-    arg_array = pd.read_csv(input_file, header=None,dtype = MapKernelUtils.get_numpy_type(dtype)).values.reshape(rows, cols)
-
+    arg_array = np.fromfile(input_file, dtype=PythonMapKernelUtils.get_numpy_type(dtype)).reshape(rows, cols)    
     match = re.search(r'def (\w+)', func)
     if match:
         try:
@@ -55,4 +51,4 @@ def apply_map_function(input_file, output_file, rows, cols, func, varName, dtype
         except Exception as e:
             print(f"Failed to execute lambda expression: {str(e)}")
 
-    pd.DataFrame(res_array).to_csv(output_file, index=False, header=False)
+    res_array.tofile(output_file)
