+++ src/runtime/local/kernels/MAP_EXTERNAL/PythonMapKernel_copy.py
@@ -20,13 +20,14 @@
 # Modifications Copyright 2022 The DAPHNE Consortium
 #
 # -------------------------------------------------------------
-import MapKernelUtils
+import PythonMapKernelUtils
 import numpy as np
 from sympy import symbols, lambdify, sympify, Symbol
 import re
 
-def apply_map_function(input_file, output_file, rows, cols, func, varName, dtype):
-    arg_array = np.fromfile(input_file, dtype=MapKernelUtils.get_numpy_type(dtype)).reshape(rows, cols)    
+def apply_map_function(arg_list, rows, cols, func, varName, dtype_arg):
+    arg_array = np.array(arg_list, dtype=PythonMapKernelUtils.get_numpy_type(dtype_arg)).reshape(rows, cols)
+    
     match = re.search(r'def (\w+)', func)
     if match:
         try:
@@ -37,7 +38,8 @@ def apply_map_function(input_file, output_file, rows, cols, func, varName, dtype
             func_name = match.groups()[0]
             func_obj = context.get(func_name)
             if func_obj:
-                res_array = np.vectorize(func_obj, otypes=[dtype])(arg_array)
+                res_array = np.vectorize(func_obj, otypes=[PythonMapKernelUtils.get_numpy_type(dtype_arg)])(arg_array)
+                return res_array.flatten().tolist()
             else:
                 print(f"Function '{func_name}' not found.")
         except Exception as e:
@@ -47,8 +49,8 @@ def apply_map_function(input_file, output_file, rows, cols, func, varName, dtype
             x = symbols(varName)
             func_expr = sympify(func.strip())
             func_lambda = lambdify(x, func_expr, modules=["numpy"])
-            res_array = np.array(func_lambda(arg_array), dtype=dtype)
+            res_array = np.array(func_lambda(arg_array), dtype=PythonMapKernelUtils.get_numpy_type(dtype_arg))
+            return res_array.flatten().tolist()
         except Exception as e:
             print(f"Failed to execute lambda expression: {str(e)}")
-
-    res_array.tofile(output_file)
+    return []
